chore(plot): rename bokeh_source.py into legend.py
